+++ dashboard_app/__init__.py
@@ -20,4 +20,4 @@
 Dashboard Application (package)
 """
 
-__version__ = (0, 19, 0, "dev", 0)
+__version__ = (0, 20, 0, "dev", 0)