+++ dashboard_app/urls.py
@@ -72,4 +72,8 @@ urlpatterns = patterns(
     url(r'^efforts/(?P<pk>[0-9]+)/$', 'testing_effort_detail'),
     url(r'^efforts/(?P<pk>[0-9]+)/update/$', 'testing_effort_update'),
     url(r'^efforts/(?P<project_identifier>[a-z0-9-]+)/\+new/$', 'testing_effort_create'),
+    url(r'^image-reports/$', 'image_report_list'),
+    url(r'^image-reports/(?P<name>[A-Za-z0-9_-]+)$', 'image_report_detail'),
+    url(r'^api/link-bug-to-testrun', 'link_bug_to_testrun'),
+    url(r'^api/unlink-bug-and-testrun', 'unlink_bug_and_testrun'),
 )